[dashboard/aria] Fix syntax error (`TSX` in `.ts` file)
