+++ app/gui/src/dashboard/components/aria/TabPanel.tsx
@@ -1,9 +1,7 @@
 /** @file Barrel re-export of `react-aria` and `react-aria-components`. */
 import { createHideableComponent } from '@react-aria/collections'
-import type { Mutable } from 'enso-common/src/utilities/data/object'
 import type { ForwardedRef } from 'react'
 import { useContext } from 'react'
-import * as aria from 'react-aria'
 import { useFocusRing } from 'react-aria'
 import {
   UNSTABLE_CollectionRendererContext as CollectionRendererContext,
@@ -13,36 +11,7 @@ import {
   type TabPanelProps,
 } from 'react-aria-components'
 import invariant from 'tiny-invariant'
-
-export * from '@react-aria/interactions'
-export { ClearPressResponder } from '@react-aria/interactions'
-export type * from '@react-types/shared'
-export * from 'react-aria'
-// @ts-expect-error The conflicting exports are props types ONLY
-export * from 'react-aria-components'
-// Resolve ambigouous star exports (`react-aria` and `react-aria-components`)
-export { I18nProvider, RouterProvider } from 'react-aria-components'
-export {
-  useTooltipTriggerState,
-  type OverlayTriggerState,
-  type TooltipTriggerState,
-} from 'react-stately'
-
-/**
- * Merges multiple props objects together.
- * Event handlers are chained, classNames are combined, and ids are deduplicated -
- * different ids will trigger a side-effect and re-render components hooked up with `useId`.
- * For all other props, the last prop object overrides all previous ones.
- *
- * The constraint is defaulted to `never` to make an explicit constraint mandatory.
- */
-export function mergeProps<Constraint extends object = never>() {
-  return <const T extends readonly (Partial<Constraint> | null | undefined)[]>(
-    ...args: T & { [K in keyof T]: Pick<T[K], keyof Constraint & keyof T[K]> }
-    // This is SAFE, as `args` is an intersection of `T` and another type.
-    // eslint-disable-next-line no-restricted-syntax
-  ) => aria.mergeProps<Mutable<T>>(...(args as T))
-}
+import { mergeProps } from './aria'
 
 /**
  * A TabPanel provides the content for a tab.
